+++ datasource/src/main/scala/astraea/spark/rasterframes/datasource/geotiff/package.scala
@@ -22,25 +22,34 @@ package astraea.spark.rasterframes.datasource
 import java.net.URI
 
 import astraea.spark.rasterframes._
-import org.apache.spark.sql.DataFrameReader
+import org.apache.spark.sql.{DataFrameReader, DataFrameWriter}
 import shapeless.tag
 import shapeless.tag.@@
 
 /**
+ * Extension methods enabled by this module.
  *
  * @since 1/16/18
  */
 package object geotiff {
-
   /** Tagged type construction for enabling type-safe extension methods for loading
    * a RasterFrame in expected form. */
   type GeoTiffRasterFrameReader = DataFrameReader @@ GeoTiffRasterFrameReaderTag
   trait GeoTiffRasterFrameReaderTag
+  /** Tagged type construction for enabling type-safe extension methods for writing
+   * a RasterFrame to a GeoTIFF. */
+  type GeoTiffRasterFrameWriter[T] = DataFrameWriter[T] @@ GeoTiffRasterFrameWriterTag
+  trait GeoTiffRasterFrameWriterTag
 
   /** Adds `geotiff` format specifier to `DataFrameReader`. */
   implicit class DataFrameReaderHasGeoTiffFormat(val reader: DataFrameReader) {
     def geotiff: GeoTiffRasterFrameReader =
-      tag[GeoTiffRasterFrameReaderTag][DataFrameReader](reader.format("geotiff"))
+      tag[GeoTiffRasterFrameReaderTag][DataFrameReader](reader.format(DefaultSource.SHORT_NAME))
+  }
+
+  implicit class DataFrameWriterHasGeoTiffFormat[T](val writer: DataFrameWriter[T]) {
+    def geotiff: GeoTiffRasterFrameWriter[T] =
+      tag[GeoTiffRasterFrameWriterTag][DataFrameWriter[T]](writer.format(DefaultSource.SHORT_NAME))
   }
 
   /** Adds `loadRF` to appropriately tagged `DataFrameReader` */