+++ docs/src/main/tut/release-notes.md
@@ -4,12 +4,12 @@
 
 ### 0.6.2
 
+* Added for for writing GeoTIFFs from RasterFrames via with DataFrameWriter.
 * _Breaking_: Renamed/moved `astraea.spark.rasterframes.functions.CellStatsAggregateFunction.Statistics` to
 `astraea.spark.rasterframes.stats.CellStatistics`.
 * _Breaking_: `HistogramAggregateFunction` now generates a `astraea.spark.rasterframes.stats.CellHistogram`.
 
   
-
 ### 0.6.1
 
 * Added support for reading striped GeoTiffs (#64).